# docs/swagger.yaml
basePath: /
definitions:
  access.CheckInput:
    properties:
      action:
        example: read
        type: string
      resource:
        example: documents
        type: string
      tenant_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      user_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
    type: object
  access.CheckResponse:
    properties:
      allowed:
        type: boolean
    type: object
  action.Action:
    properties:
      app_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      metadata:
        type: object
      name:
        example: read
        type: string
      updated_at:
        type: string
    type: object
  action.CreateInput:
    properties:
      metadata:
        type: object
      name:
        example: read
        type: string
    type: object
  action.ListResponse:
    properties:
      actions:
        items:
          $ref: '#/definitions/action.Action'
        type: array
    type: object
  action.UpdateInput:
    properties:
      metadata:
        type: object
      name:
        example: read
        type: string
    type: object
...
  apikey.APIKey:
    properties:
      app_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      key:
        description: plaintext, only on create
        type: string
      metadata:
        type: object
      name:
        type: string
      revoked_at:
        type: string
    type: object
  apikey.CreateInput:
    properties:
      metadata:
        type: object
      name:
        type: string
    type: object
  apikey.ListResponse:
    properties:
      api_keys:
        items:
          $ref: '#/definitions/apikey.APIKey'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: OpenAura
  description: |-
    Open-source user, role, and access management service.
    API version is selected via the X-API-Version header (not the URL path).
    App routes require an app API key (X-API-Key). Admin routes live under /admin and require an admin API key.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenAura API
  version: "1.0"
paths:
  /access/check:
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Access check request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/access.CheckInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/access.CheckResponse'
      summary: Check access
      tags:
      - access
  /actions:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/action.ListResponse'
      summary: List actions
      tags:
      - actions
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Action to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/action.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/action.Action'
      summary: Create action
      tags:
      - actions
  /actions/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Action ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete action
      tags:
      - actions
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Action ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/action.Action'
      summary: Get action
      tags:
      - actions
    patch:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Action ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/action.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/action.Action'
      summary: Update action
      tags:
      - actions
...
  /admin/apps/{id}/api_keys:
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: App ID
        in: path
        name: id
        required: true
        type: string
      - description: API key to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apikey.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apikey.APIKey'
      summary: Create app API key (admin)
      tags:
      - admin-apps
  /api_keys:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apikey.ListResponse'
      summary: List app API keys
      tags:
      - api_keys
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: API key to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apikey.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apikey.APIKey'
      summary: Create app API key
      tags:
      - api_keys
  /api_keys/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Revoke app API key
      tags:
      - api_keys
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apikey.APIKey'
      summary: Get app API key
      tags:
      - api_keys
...
schemes:
- http
swagger: "2.0"
...